actions/InjectFlags: Add config option for linker flags
This adds the configuration options `CFLAGS_LINKER` and `CXXFLAGS_LINKER`
to the InjectFlags action. The flags specified in those will only be
appended if the compiler tool will run the linking stage.

For calls that don't run the linking stage (such as those that include
the `-c`, `-S`, `-E`, etc options), these flags won't be appended.